requirements.txt
openai==1.75.0
python-dotenv==1.0.1
langgraph==0.6.11
typing-extensions==4.13.2